+++ lib/qr.dart
@@ -1,4 +1,5 @@
 export 'src/bit_buffer.dart';
 export 'src/error_correct_level.dart';
 export 'src/input_too_long_exception.dart';
-export 'src/qr_code.dart' hide qrModules;
+export 'src/qr_code.dart';
+export 'src/qr_image.dart';