+++ test/qr_code_test.dart
@@ -4,6 +4,7 @@ import 'dart:typed_data';
 
 import 'package:qr/src/error_correct_level.dart';
 import 'package:qr/src/qr_code.dart';
+import 'package:qr/src/qr_image.dart';
 import 'package:test/test.dart';
 
 import 'qr_code_test_data.dart';
@@ -13,12 +14,11 @@ void main() {
   test('simple', () {
     for (var typeNumber = 1; typeNumber <= 40; typeNumber++) {
       for (var quality in QrErrorCorrectLevel.levels) {
-        final qr = QrCode(typeNumber, quality)
-          ..addData('shanna!')
-          ..make();
-        for (var i = 0; i < qrModules(qr).length; i++) {
+        final qr = QrImage(QrCode(typeNumber, quality)..addData('shanna!'));
+        final modules = qr.qrModules;
+        for (var i = 0; i < modules.length; i++) {
           expect(
-            _encodeBoolListToString(qrModules(qr)[i]),
+            _encodeBoolListToString(modules[i]),
             qrCodeTestData[typeNumber.toString()][quality.toString()][i],
           );
         }
@@ -28,11 +28,12 @@ void main() {
 
   test('fromData', () {
     for (var quality in QrErrorCorrectLevel.levels) {
-      final qr = QrCode.fromData(data: 'shanna!', errorCorrectLevel: quality)
-        ..make();
-      for (var i = 0; i < qrModules(qr).length; i++) {
+      final qr =
+          QrImage(QrCode.fromData(data: 'shanna!', errorCorrectLevel: quality));
+      final modules = qr.qrModules;
+      for (var i = 0; i < modules.length; i++) {
         expect(
-          _encodeBoolListToString(qrModules(qr)[i]),
+          _encodeBoolListToString(modules[i]),
           qrCodeTestData['1'][quality.toString()][i],
         );
       }
@@ -41,13 +42,13 @@ void main() {
 
   test('fromUint8List', () {
     for (var quality in QrErrorCorrectLevel.levels) {
-      final qr = QrCode.fromUint8List(
+      final qr = QrImage(QrCode.fromUint8List(
           data: Uint8List.fromList([115, 104, 97, 110, 110, 97, 33]),
-          errorCorrectLevel: quality)
-        ..make();
-      for (var i = 0; i < qrModules(qr).length; i++) {
+          errorCorrectLevel: quality));
+      final modules = qr.qrModules;
+      for (var i = 0; i < modules.length; i++) {
         expect(
-          _encodeBoolListToString(qrModules(qr)[i]),
+          _encodeBoolListToString(modules[i]),
           qrCodeTestData['1'][quality.toString()][i],
         );
       }
@@ -56,12 +57,12 @@ void main() {
 
   test('WHEN mask pattern is provided, SHOULD make a masked QR Code', () {
     for (var mask = 0; mask <= 7; mask++) {
-      final qr = QrCode(1, QrErrorCorrectLevel.L)
-        ..addData('shanna!')
-        ..make(mask);
-      for (var i = 0; i < qrModules(qr).length; i++) {
+      final qr = QrImage.withMaskPattern(
+          QrCode(1, QrErrorCorrectLevel.L)..addData('shanna!'), mask);
+      final modules = qr.qrModules;
+      for (var i = 0; i < modules.length; i++) {
         expect(
-          _encodeBoolListToString(qrModules(qr)[i]),
+          _encodeBoolListToString(modules[i]),
           qrCodeTestDataWithMask[mask.toString()][i],
         );
       }
@@ -73,9 +74,8 @@ void main() {
       SHOULD throw an AssertionError
   ''', () {
     expect(() {
-      QrCode(1, QrErrorCorrectLevel.L)
-        ..addData('shanna!')
-        ..make(-1);
+      QrImage.withMaskPattern(
+          QrCode(1, QrErrorCorrectLevel.L)..addData('shanna!'), -1);
     }, throwsA(isA<AssertionError>()));
   });
 
@@ -84,9 +84,8 @@ void main() {
       SHOULD throw an AssertionError
   ''', () {
     expect(() {
-      QrCode(1, QrErrorCorrectLevel.L)
-        ..addData('shanna!')
-        ..make(8);
+      QrImage.withMaskPattern(
+          QrCode(1, QrErrorCorrectLevel.L)..addData('shanna!'), 8);
     }, throwsA(isA<AssertionError>()));
   });
 }